+++ src/methods/delete/delete.ts
@@ -1,14 +1,6 @@
-import { TServerMethodWithId, TUser } from './types.js';
-import { findUserById } from './findById.js';
-import { users } from './data/users.js';
-
-const remove = (user: TUser): Promise<void> => {
-  return new Promise((resolve) => {
-    const userIndex = users.indexOf(user);
-    users.splice(userIndex, 1);
-    resolve();
-  });
-};
+import { TServerMethodWithId } from '../../types.js';
+import { findUserById } from '../findById.js';
+import { remove } from './removeInDB.js';
 
 export const deleteUser: TServerMethodWithId = async (
   _request,
