+++ src/methods/getRequestBody.ts
@@ -1,4 +1,4 @@
-import { TGetRequestBody } from './types.js';
+import { TGetRequestBody } from '../types.js';
 
 export const getRequestBody: TGetRequestBody = (request) => {
   return new Promise((resolve, reject) => {
